requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiologger==0.7.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
exceptiongroup==1.2.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
iniconfig==2.0.0
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
yarl==1.9.4
